+++ src/main/java/nl/clockwork/ebms/dao/spring/mysql/EbMSDAOImpl.java
@@ -61,7 +61,7 @@ public class EbMSDAOImpl extends AbstractEbMSDAO
 	}
 
 	@Override
-	public String getMessageIdsQuery(String messageContextFilter, EbMSMessageStatus status, long maxNr)
+	public String getMessageIdsQuery(String messageContextFilter, EbMSMessageStatus status, int maxNr)
 	{
 		return "select message_id" +
 		" from ebms_message" +
